Use Thor instead of OptionParser
Also improves the `new` command to support any defined rules. Created
feature note files will include all supported keys.

Using Thor instead of OptionParser breaks the previous behavior of the
CLI. Using

    renogen SOME_VERSION

will not work anymore. You either have to use the `generate` subcommand
explicitely or make it clear that `SOME_VERSION` is an argument:

    renogen -- SOME_VERSION

Sorry for that breaking change, but the benefits of using Thor are huge
(for instance, existing files won't be overwritten, the user is asked
what to do instead).